Split Haiku code into its own crate
