+++ Cargo.toml
@@ -31,7 +31,8 @@ wasm-bindgen = "0.2.70"
 wasm-bindgen-test = "0.3"
 
 [target.'cfg(target_os = "haiku")'.dependencies]
-cxx = "1.0.34"
+iana-time-zone-haiku = { version = "0.1.0", path = "haiku" }
 
-[target.'cfg(target_os = "haiku")'.build-dependencies]
-cxx-build = "1.0.34"
+[workspace]
+members = [".", "haiku"]
+default-members = ["."]
