+++ haiku/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "iana-time-zone-haiku"
+description = "iana-time-zone support crate for Haiku OS"
+version = "0.1.0"
+authors = ["René Kijewski <crates.io@k6i.de>"]
+repository = "https://github.com/strawlab/iana-time-zone"
+license = "MIT OR Apache-2.0"
+readme = "README.md"
+edition = "2018"
+
+[dependencies]
+cxx = "1.0.34"
+
+[build-dependencies]
+cxx-build = "1.0.34"
